Now checking to make sure currentIndex is valid before assuming it is
In the rare case that the user has manually removed some values from the preferredLineLengthValues array and the currentIndex is saved at an index that is no longer valid, the package will stop working.  I'm now checking for this.  In the case that we don't have a valid currentIndex value, just set it to 0.